[TCcnt] Fixed TCcountIds warnings and made int_matrix indices unsigned I  used casts to fix warnings regarding: A NLUT elems as stored in int, B  when size_t from TCcountIds -> unsigned int var as unsigned is enough

+++ src/libsac2c/tree/tree_basic.c
@@ -216,7 +216,7 @@ TBfreeCudaIndex (cuda_index_t *index)
 /*--------------------------------------------------------------------------*/
 
 cuda_access_info_t *
-TBmakeCudaAccessInfo (node *array, node *arrayshp, int dim, int cuwldim, int nestlevel)
+TBmakeCudaAccessInfo (node *array, node *arrayshp, int dim, size_t cuwldim, size_t nestlevel)
 {
     cuda_access_info_t *info;
     int i;
